Break the tests apart into topics
