+++ spec/integration/class_enhancement_spec.rb
@@ -1,0 +1,64 @@
+require "dry-monads"
+require "dry/result_matcher/either_matcher"
+
+RSpec.describe "Class enhancement with Dry::ResultMatcher.for" do
+  let(:operation) {
+    Class.new do
+      include Dry::ResultMatcher.for(:call, with: Dry::ResultMatcher::EitherMatcher)
+
+      def call(bool)
+        bool ? Dry::Monads::Right("a success") : Dry::Monads::Left("a failure")
+      end
+    end.new
+  }
+
+  describe "match blocks" do
+    subject(:match) {
+      operation.call(input) do |m|
+        m.success do |v|
+          "Matched success: #{v}"
+        end
+
+        m.failure do |v|
+          "Matched failure: #{v}"
+        end
+      end
+    }
+
+    context "successful result" do
+      let(:input) { true }
+
+      it "matches on success" do
+        expect(match).to eq "Matched success: a success"
+      end
+    end
+
+    context "failed result" do
+      let(:input) { false }
+
+      it "matches on failure" do
+        expect(match).to eq "Matched failure: a failure"
+      end
+    end
+  end
+
+  describe "without match blocks" do
+    subject(:result) { operation.call(input) }
+
+    context "successful result" do
+      let(:input) { true }
+
+      it "returns the result" do
+        expect(result).to eq Dry::Monads::Right("a success")
+      end
+    end
+
+    context "failed result" do
+      let(:input) { false }
+
+      it "returns the result" do
+        expect(result).to eq Dry::Monads::Left("a failure")
+      end
+    end
+  end
+end
